comfortingBotDev/requirements.txt
Flask == 1.1.1
gunicorn == 20.0.4
line-bot-sdk == 1.15.0
APScheduler == 3.6.1